Initialize default_date_column option with the rest of the config
