+++ lib/money_with_date.rb
@@ -15,6 +15,7 @@ require_relative "money_with_date/rates_store/memory"
 # :nocov:
 ::Money.date_determines_equality = false
 ::Money.default_date = ::Date.respond_to?(:current) ? -> { ::Date.current } : -> { ::Date.today }
+::Money.default_date_column = :created_at
 
 require "money_with_date/railtie" if defined?(::Rails::Railtie) && defined?(::MoneyRails)
 # :nocov:
