--- lib/money_with_date/active_record/class_methods.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module MoneyWithDate
-  module ActiveRecord
-    module ClassMethods
-      attr_accessor :default_date_column
-    end
-  end
-end
